+++ CHANGELOG.md
@@ -5,6 +5,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.12.2] - 2020-29-06
+### Changed
+- "FanMode", "ChargeLimit" to "SetFanMode", "SetChargeLimit"
+ 
+### Added
+- Dbus signals "FanModeChanged", "ChargeLimitChanged"
+- Dbus methods "GetFanMode", "GetChargeLimit"
+- Support for ROG Strix G712
+
+## [0.12.0] - 2020-26-06
 ### Changed
 - Add modes for FX531 LEDs
 - Change where USB reset is called